use the new yml config file for setting the EA branch

+++ script/clone-early-access.js
@@ -10,7 +10,8 @@
 // [end-readme]
 
 require('dotenv').config()
-const { GITHUB_DOCUBOT_REPO_PAT, EARLY_ACCESS_ENABLED } = process.env
+// const { GITHUB_DOCUBOT_REPO_PAT, EARLY_ACCESS_ENABLED, HEROKU_PRODUCTION_APP } = process.env
+const { GITHUB_DOCUBOT_REPO_PAT, HEROKU_PRODUCTION_APP } = process.env
 
 // TODO...
 // // Exit if early access is not enabled
@@ -29,28 +30,32 @@ const { execSync } = require('child_process')
 const rimraf = require('rimraf').sync
 const fs = require('fs')
 const path = require('path')
+const yaml = require('js-yaml')
+const eaConfig = yaml.load(fs.readFileSync(path.join(process.cwd(), 'ea-config.yml'), 'utf8'))
 
 // Early Access details
+const earlyAccessOwner = 'docs'
 const earlyAccessDir = 'early-access-test'
-const earlyAccessRepo = `https://${GITHUB_DOCUBOT_REPO_PAT}@github.com/docs/${earlyAccessDir}`
+const earlyAccessRepo = `https://${GITHUB_DOCUBOT_REPO_PAT}@github.com/${earlyAccessOwner}/${earlyAccessDir}`
 const earlyAccessContentDir = path.join(process.cwd(), 'content', earlyAccessDir)
 
-// TODO...
-// // Look for a branch in early-access that matches the current docs branch;
-// // otherwise fall back to `main`
-// const docsBranch = execSync('git branch --show-current').toString().trim()
-//
-// const earlyAccessBranch = execSync(`git ls-remote --heads ${earlyAccessRepo} ${docsBranch}`).toString()
-//   ? docsBranch
-//   : 'main'
-const earlyAccessBranch = 'main'
+// production vs. staging environment
+const environment = HEROKU_PRODUCTION_APP ? 'production' : 'staging'
+const earlyAccessBranch = HEROKU_PRODUCTION_APP ? eaConfig.EA_PRODUCTION_BRANCH : eaConfig.EA_STAGING_BRANCH
+
+// confirm that the branch exists in the remote
+const doesBranchExist = execSync(`git ls-remote --heads ${earlyAccessRepo} ${earlyAccessBranch}`).toString()
+if (!doesBranchExist) {
+  console.log(`The branch '${earlyAccessBranch}' was not found in ${earlyAccessOwner}/${earlyAccessDir}. Exiting!`)
+  process.exit(0)
+}
 
 // Remove any dir that may pre-exist
 rimraf(earlyAccessContentDir)
 
 // Clone the repo
 execSync(`git clone --single-branch --branch ${earlyAccessBranch} ${earlyAccessRepo} ${earlyAccessContentDir}`)
-console.log(`Branch: ${earlyAccessBranch}`)
+console.log(`Using early-access ${environment} branch: '${earlyAccessBranch}'`)
 
 // Remove the .git dir
 rimraf(`${earlyAccessContentDir}/.git`)
